requirements.txt
aiohttp==3.6.3
appnope==0.1.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
discord.py==1.5.1
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
multidict==4.7.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-dotenv==0.15.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
yarl==1.5.1
